+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Fixed #PAC-280: Extension library requires `etc` configuration directory
 * Fixed #PAC-206: Prevent finder mappings of different libraries to be overwritten
 * Fixed #PAC-318: Serialization exception when using a custom database configuration
+* Add #PAC-486: Add Getter and Setter for `--log-file` commandline parameter 
 
 ## Features